third_party/WebKit/LayoutTests/fast/scrolling/scrollbar-prevent-default-expected.txt

This tests that scrollbars always receive events even when there is an element underneath the scrollbar which swallows the event. In this case an event handler is added to the window which swallows the event but this may also happen with elements underneath of overlay scrollbars.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS receivedMousedownEvent is true
PASS window.scrollY is 0
PASS receivedMousedownEvent is true
PASS subframe.scrollTop is 0
PASS successfullyParsed is true

TEST COMPLETE
